pyproject.toml
[tool.poetry]
name = "deepaipng"
version = "1.1.1"
description = "Processes images with DeepAI and PIL by resizing, cropping to fit specified dimensions, making the background transparent"
authors = ["DJ Magar <dj@deepai.org>"]
license = "ISC"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
pillow = "^10.2.0"
httpx = ">=0.27.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
twine = [
    {version = "^5.0.0", python = ">=3.10"}
]
pytest = [
    {version = "^8.2.0", python = ">=3.10"}
]
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
